+++ db/migrate/20200723205531_create_events.rb
@@ -1,7 +1,9 @@
 class CreateEvents < ActiveRecord::Migration[6.0]
   def change
     create_table :events do |t|
+      t.string :title
       t.text :description
+      t.date :date
       t.references :user, null: false, foreign_key: true
 
       t.timestamps