+++ CHANGELOG.md
@@ -14,3 +14,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - View user events.
 - View all events.
 - Attend/Unattend events.
+- Separate attended from unattended events.
+- Separate upcoming from past events.